+++ gridsome/lib/app/App.js
@@ -1,9 +1,9 @@
 const path = require('path')
 const isUrl = require('is-url')
-const Pages = require('./pages')
 const Codegen = require('./codegen')
 const autoBind = require('auto-bind')
 const hirestime = require('hirestime')
+const Pages = require('../pages/pages')
 const BaseStore = require('./BaseStore')
 const PluginAPI = require('./PluginAPI')
 const { execute, graphql } = require('graphql')
