+++ gridsome/lib/pages/__tests__/__fixtures__/DefaultPage.vue
