+++ gridsome/lib/pages/__tests__/__fixtures__/DefaultTemplate.vue
